SME-MK mode support
SME-MK mode changes the way the PTE is used to specify encryption. When
SME-MK mode is enabled, the PTE now identifies the index of the encryption
key that is used.

TSME detection is not possible in SME-MK mode. Add support to detect this
mode and issue an error message during module initialization.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>